python-ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.40.0
mysql-connector-python==8.2.0
PyPDF2==3.0.1
pdf2image==1.16.3
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
